RF: Move check_dates() out of plugins
Towards a fix for #4519
